Move app specific actions to apps group
